+++ plugins/guests/bsd/guest.rb
@@ -2,7 +2,7 @@ module VagrantPlugins
   module GuestBSD
     class Guest < Vagrant.plugin("2", :guest)
       def detect?(machine)
-        machine.communicate.test("uname -s | grep -i 'Darwin|BSD'")
+        machine.communicate.test("uname -s | grep -i 'BSD'")
       end
     end
   end