+++ plugins/guests/darwin/cap/halt.rb
@@ -7,7 +7,7 @@ module VagrantPlugins
             # Darwin does not support the `-p` option like the rest of the
             # BSD-based guests, so it needs its own cap.
             machine.communicate.sudo("/sbin/shutdown -h now")
-          rescue IOError
+          rescue IOError, Vagrant::Errors::SSHDisconnected
             # Do nothing because SSH connection closed and it probably
             # means the VM just shut down really fast.
           end