+++ plugins/guests/dragonflybsd/plugin.rb
@@ -1,0 +1,15 @@
+require "vagrant"
+
+module VagrantPlugins
+  module GuestDragonFlyBSD
+    class Plugin < Vagrant.plugin("2")
+      name "DragonFly BSD guest"
+      description "DragonFly BSD guest support."
+
+      guest(:dragonflybsd, :bsd) do
+        require_relative "guest"
+        Guest
+      end
+    end
+  end
+end