+++ plugins/guests/solaris11/cap/rsync.rb
@@ -1,10 +1,10 @@
+require_relative "../../../synced_folders/rsync/default_unix_cap"
+
 module VagrantPlugins
   module GuestSolaris11
     module Cap
       class RSync
-        def self.rsync_installed(machine)
-          machine.communicate.test("which rsync")
-        end
+        extend VagrantPlugins::SyncedFolderRSync::DefaultUnixCap
 
         def self.rsync_command(machine)
           "#{machine.config.solaris11.suexec_cmd} rsync"
@@ -17,10 +17,11 @@ module VagrantPlugins
         end
 
         def self.rsync_post(machine, opts)
+          if opts.key?(:chown) && !opts[:chown]
+            return
+          end
           suexec_cmd = machine.config.solaris11.suexec_cmd
-          machine.communicate.execute(
-            "#{suexec_cmd} '#{opts[:guestpath]}' '(' ! -user #{opts[:owner]} -or ! -group #{opts[:group]} ')' -print0 | " +
-              "xargs -0 chown #{opts[:owner]}:#{opts[:group]}")
+          machine.communicate.execute("#{suexec_cmd} #{build_rsync_chown(opts)}")
         end
       end
     end