+++ test/unit/plugins/guests/darwin/cap/halt_test.rb
@@ -32,5 +32,12 @@ describe "VagrantPlugins::GuestDarwin::Cap::Halt" do
         cap.halt(machine)
       }.to_not raise_error
     end
+
+    it "ignores a Vagrant::Errors::SSHDisconnected" do
+      comm.stub_command("/sbin/shutdown -h now", raise: Vagrant::Errors::SSHDisconnected)
+      expect {
+        cap.halt(machine)
+      }.to_not raise_error
+    end
   end
 end